test/requirements.txt
dtw-python==1.5.3
numpy==2.0.2
scipy==1.13.1
